adding all controllers

+++ src/main/java/com/binaryBaaje/ShiftSnap/controller/UserController.java
@@ -4,11 +4,13 @@ import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -30,8 +32,8 @@ public class UserController {
 	
 	
 	@GetMapping("/users")
-	public List<User> retrieveAllUser(){
-		return userRepository.findAll();
+	public ResponseEntity<List<User>> retrieveAllUser(){
+		return new ResponseEntity<>(userRepository.findAll(), HttpStatus.OK);
 	}
 	
 	@GetMapping("/users/{userId}")
@@ -59,24 +61,26 @@ public class UserController {
 		
 	}
 	
-	@DeleteMapping("/users/{userId}")
-	public void deleteUser(@PathVariable Long userId) {
-		userRepository.deleteById(userId);
-		
-	}
-	
-	@PostMapping("/users/{userId}")
-	public ResponseEntity<Object> upateUser(@Valid @RequestBody User user) {
-		User savedUser =userRepository.save(user);
-		
-		URI location = ServletUriComponentsBuilder
-				.fromCurrentRequest()
-				.path("/{userId}")
-				.buildAndExpand(savedUser.getUserId())
-				.toUri();
-		return ResponseEntity.created(location ).build();
-		
-	}
-	
 
+
+	@DeleteMapping("/users/{userId}")
+	public ResponseEntity<String> deleteUser(@PathVariable Long userId) {
+		Optional<User> user = userRepository.findById(userId);
+		if(user.isPresent()){
+			userRepository.deleteById(userId);
+			return new ResponseEntity<>("Deleted", HttpStatus.OK);
+		}
+		return new ResponseEntity<>("User Not Found", HttpStatus.NOT_FOUND);
+		
+	}
+
+	@PutMapping("/users/{userId}")
+	public ResponseEntity<Object> updateUser(@Valid @RequestBody User user, @PathVariable Long userId) {
+		if(userRepository.existsById(userId)){
+			user.setUserId(userId);
+			userRepository.save(user);
+			return new ResponseEntity<>("User Updated Successfully", HttpStatus.OK);
+		}
+		return new ResponseEntity<>("User Not Found", HttpStatus.NOT_FOUND);
+	}
 }
